requirements.txt
chainlit
openplugins
flask
flask-cors
waitress
baostock
akshare
colorlog
python-dotenv

#stockasisstant
flask
flask-cors
waitress
baostock
akshare
colorlog
#openaifunctioncall
openai
tenacity 
pyyaml
requests